Fix for the non-NCCL code path

+++ legate/pandas/core/sort.py
@@ -202,7 +202,7 @@ class Sorter(object):
 
         result_storage = plan.promote_output_storage(result_storage)
         rt.register_external_weighted_partition(
-            result_storage.default_ipart, counts.cast(ty.int32)
+            result_storage.default_ipart, counts.cast(ty.int64)
         )
         del plan
